pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "investigator-ai"
version = "0.1.0"
description = "A comprehensive fraud investigation tool that aggregates data from legitimate public sources"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "InvestigatorAI Team"}
]
license = {file = "LICENSE"}
keywords = ["fraud", "investigation", "data", "analysis", "compliance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Legal Industry", 
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    # Core data collection and analysis
    "requests==2.32.4",
    "pandas==2.3.1",
    "numpy==2.3.2",
    "kaggle==1.7.4.5",
    "python-dotenv==1.1.1",
    "PyPDF2==3.0.1",
    "openpyxl==3.1.5",
    # AI/ML and LangChain dependencies for enhanced notebook
    "langchain-openai==0.3.28",
    "langchain-core==0.3.72", 
    "langgraph==0.6.1",
    "langchain-qdrant>=0.2.0",
    "pydantic==2.11.7",
    "qdrant-client==1.15.0",
    "ragas==0.3.0",
    "datasets==2.21.0",
    "typing-extensions==4.14.1",
    "PyMuPDF==1.24.1",
    # FastAPI dependencies
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]
notebook = [
    "jupyter>=1.0.0",
    "jupyterlab>=3.6.0",
    "notebook>=6.4.0",
    "ipykernel>=6.15.0",
    "ipython>=8.0.0",
    "ipywidgets>=8.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "plotly>=5.0.0",
]
full = [
    "investigator-ai[dev,notebook]",
]

[project.urls]
Homepage = "https://github.com/ai-makerspace/InvestigatorAI"
Repository = "https://github.com/ai-makerspace/InvestigatorAI"
Issues = "https://github.com/ai-makerspace/InvestigatorAI/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["investigator_ai*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=investigator_ai --cov-report=html --cov-report=term-missing"
